Implement Wave 2 local seam stabilization
